fix the jthread data member destruction order bug
We were relying too much on the fact that when jthread is used as a data
member, it will be auto-joined when its containing object is destroyed.
But given the data member destruction order, it means that all data
member declared below the jthread will be be destroyed before the
jthread is auto-joined;

Our problem is that we expect the `prot_th` thread to
call the stop() method on the AsyncMsgScanner, which will cancel the
blocked IO and wake up and join the scanner `tup` thread. But since the amsc data
member is declared below the `prot_ht` jthread in the Session class, the
dtor of AsyncMsgScanner is called before the erase-remove idiom in the
cleanUp method blocks the gc thread on joining the prot_th;
So we see a premature std::thread object dtor called before its thread (the `tup` std::thread)
is joined.

The solution we made is to make `tup` also a jthread; So the gc thread
will first block on joining the amsc `tup` thread (which is in turn
block in a `sleep(1)` waiting for pthread_cancel issued from the prot_th).
i.e the destruction of AsyncMsgScanner itself will wait for tup to be joined.